refactor(puniyu_utils):重构配置模块并优化路径处理 - 将配置相关代码从 utils/config 移动到 config 目录 - 新增 path 模块，用于处理配置文件路径- 删除 number 模块 - 更新相关模块引用和路径处理方式

+++ puniyu_utils/src/config/yaml.rs
@@ -1,68 +1,68 @@
 use super::utils::delete_nested_node;
 use serde::{Serialize, de::DeserializeOwned};
+use serde_yaml::{Value, from_str, to_string};
 use std::{error::Error, fs, path::Path};
-use toml::{from_str, to_string_pretty};
 
-/// 读取toml配置文件
+/// 读取yaml配置文件
 ///
 /// # 参数
 ///
 /// `path` 配置文件路径
 /// `name` 配置文件名, 不包含后缀
-pub fn read_config<D>(path: &Path, name: &str) -> Result<D, Box<dyn Error>>
+pub fn read_config<C>(path: &Path, name: &str) -> Result<C, Box<dyn Error>>
 where
-    D: DeserializeOwned,
+    C: DeserializeOwned,
 {
-    let full_path = path.join(format!("{}.toml", name));
+    let full_path = path.join(format!("{}.yaml", name));
     let config_str = fs::read_to_string(full_path)?;
-    let config: D = from_str(&config_str)?;
+    let config: C = from_str(&config_str)?;
     Ok(config)
 }
-/// 写入toml配置文件
+
+/// 写入yaml配置文件
 ///
 /// # 参数
 ///
 /// `path` 配置文件路径
 /// `name` 配置文件名, 不包含后缀
+/// `config` 配置数据
 pub fn write_config<C>(path: &Path, name: &str, config: &C) -> Result<(), Box<dyn Error>>
 where
     C: Serialize,
 {
-    let full_path = path.join(format!("{}.toml", name));
-    if let Some(parent) = path.parent() {
+    let full_path = path.join(format!("{}.yaml", name));
+    if !full_path.exists() {
+        return Ok(());
+    }
+    if let Some(parent) = full_path.parent() {
         fs::create_dir_all(parent)?;
     }
-    let config_str = to_string_pretty(config)?;
+    let config_str = to_string(config)?;
     fs::write(full_path, config_str)?;
     Ok(())
 }
 
-/// 更新toml配置文件
+/// 更新yaml配置文件
 ///
 /// # 参数
 ///
 /// `path` 配置文件路径
 /// `name` 配置文件名, 不包含后缀
+/// `updater` 更新函数, 用于更新配置
 ///
 /// # 示例
-///
 /// ```
 /// use serde::{Deserialize, Serialize};
 /// use std::path::Path;
-/// use puniyu_utils::utils::config::toml::update_config;
-///
+/// use puniyu_utils::config::yaml::update_config;
 /// #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 /// pub struct Config {
-///     field1: String,
-///     field2: u32,
+/// pub field: String,
 /// }
-///
-/// update_config::<Config>(Path::new("config"), "config", |config| {
-///     config.field1 = "new value".to_string();
-///     config.field2 = 42;
-/// })?;
-/// ```
-///
+/// let config_path = Path::new("./config");
+/// update_config::<Config>(config_path, "bot", |config| {
+/// config.field = "loli".to_string()
+/// }).unwrap();
 pub fn update_config<C>(
     path: &Path,
     name: &str,
@@ -71,7 +71,10 @@ pub fn update_config<C>(
 where
     C: Serialize + DeserializeOwned + Default,
 {
-    let full_path = path.join(format!("{}.toml", name));
+    let full_path = path.join(format!("{}.yaml", name));
+    if !full_path.exists() {
+        return Ok(());
+    }
 
     let mut config = read_config(path, name).unwrap_or_else(|_| C::default());
 
@@ -80,12 +83,12 @@ where
     if let Some(parent) = full_path.parent() {
         fs::create_dir_all(parent)?;
     }
-    let config_str = to_string_pretty(&config)?;
+    let config_str = to_string(&config)?;
     fs::write(full_path, config_str)?;
     Ok(())
 }
 
-/// 删除toml配置文件中的指定节点
+/// 删除yaml配置文件中的指定节点
 ///
 /// # 参数
 ///
@@ -94,36 +97,37 @@ where
 /// `node_path` 要删除的节点路径，支持嵌套节点，如 "parent.child"
 ///
 /// # 示例
+///
 /// ```
 /// use std::path::Path;
-/// use puniyu_utils::utils::config::toml::delete_config;
+/// use puniyu_utils::config::yaml::delete_config;
 /// let config_path = Path::new("./config");
-/// delete_config(config_path, "bot", "field").unwrap();
-/// ```
+/// delete_config(config_path, "bot", "logger_level").unwrap(); // 删除bot配置中的logger_level节点
+///```
 pub fn delete_config(path: &Path, name: &str, node_path: &str) -> Result<(), Box<dyn Error>> {
-    let full_path = path.join(format!("{}.toml", name));
-
+    let full_path = path.join(format!("{}.yaml", name));
     if !full_path.exists() {
         return Ok(());
     }
 
     let config_str = fs::read_to_string(&full_path)?;
-    let mut toml_value: toml::Value = from_str(&config_str)?;
+    let mut yaml_value: Value = from_str(&config_str)?;
+
     let node_keys: Vec<&str> = node_path.split('.').collect();
 
     delete_nested_node(
-        &mut toml_value,
+        &mut yaml_value,
         &node_keys,
         |value, key| {
-            if let toml::Value::Table(table) = value {
-                table.get_mut(key)
+            if let Value::Mapping(mapping) = value {
+                mapping.get_mut(key)
             } else {
                 None
             }
         },
         |value, key| {
-            if let toml::Value::Table(table) = value {
-                table.remove(key);
+            if let Value::Mapping(mapping) = value {
+                mapping.remove(key);
             }
         },
     )?;
@@ -131,7 +135,7 @@ pub fn delete_config(path: &Path, name: &str, node_path: &str) -> Result<(), Box
     if let Some(parent) = full_path.parent() {
         fs::create_dir_all(parent)?;
     }
-    let config_str = to_string_pretty(&toml_value)?;
+    let config_str = to_string(&yaml_value)?;
     fs::write(full_path, config_str)?;
 
     Ok(())
